move to monk shared constants for increases
